Conflicts:

	src/mesa/drivers/dri/common/dri_bufmgr.c
	src/mesa/drivers/dri/common/dri_drmpool.c
	src/mesa/drivers/dri/i915tex/intel_batchpool.c
	src/mesa/drivers/dri/i915tex/intel_buffer_objects.c
	src/mesa/drivers/dri/i915tex/intel_regions.c
	src/mesa/drivers/dri/i915tex/intel_screen.c
	src/mesa/drivers/dri/i915tex/intel_screen.h

+++ progs/demos/arbfslight.c
@@ -62,7 +62,7 @@ static PFNGLUNIFORM3FVARBPROC glUniform4fvARB = NULL;
 
 static void normalize (GLfloat *dst, const GLfloat *src)
 {
-   GLfloat len = sqrtf (src[0] * src[0] + src[1] * src[1] + src[2] * src[2]);
+   GLfloat len = sqrt (src[0] * src[0] + src[1] * src[1] + src[2] * src[2]);
    dst[0] = src[0] / len;
    dst[1] = src[1] / len;
    dst[2] = src[2] / len;
